+++ changelog.md
@@ -5,7 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
-- Change all group membership test utilities to be functions instead of variables. You now have to call `is_<group_id>()` instead of `is_<group_id>`.
+- Change all group membership check utilities to be functions instead of variables. You now have to call `is_<group_id>()` instead of `is_<group_id>`.
+- Cache the result of `is_<group_id>()` group membership check utilities.
 - Make URLs required on all platforms.
 
 ## [1.7.0 (2024-12-02)](https://github.com/kdeldycke/extra-platforms/compare/v1.6.0...v1.7.0)