+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Change all group membership check utilities to be functions instead of variables. You now have to call `is_<group_id>()` instead of `is_<group_id>`.
 - Cache the result of `is_<group_id>()` group membership check utilities.
+- Do not call all detection heuristics on module import. Instead, call them lazily when needed.
 - Make URLs required on all platforms.
 
 ## [1.7.0 (2024-12-02)](https://github.com/kdeldycke/extra-platforms/compare/v1.6.0...v1.7.0)